+++ test/xcop.test.js
@@ -19,6 +19,7 @@ const PACKS = [
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'xpath-validator-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'xpath-format-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'axis-packs')),
+  ...allFilesFrom(path.resolve(__dirname, 'resources', 'namespace-packs')),
 ]
 
 /**